MissileScript Commit
Missile script is now in working condition, we need to add model and sound.

We have a weapon, that once fired, will ping nearest enemy, deem it a target and start homing towards it. If in case of destroyed target with missiles still targeting it, missiles will seek new target to home in to. Glorius!