turretRotation fix commit.
Okay I fixed the turret rotation, forever. It now uses constant as local position from the main ship as reference, so it cannot be wrong anymore.

I had to clone Kristian's turret, put it in same place and angles as best as I could and deactivated the mirrored turret, so the base values are same in both turrets.